requirements.txt
python-dotenv
livekit-agents
openai
deepgram-sdk
faster-whisper
cartesia
silero
